Refactor: Moving and renaming things around for better visibility.
